% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs_countries.R
\docType{data}
\name{ifs_countries}
\alias{ifs_countries}
\title{ifs_countries
A data frame containing the name of 248 countries and its ISO 3166-1 alpha-2 codes
One can use this data frame to search for the correct ISO code of the country he wants do download data using \code{\link{ifs_data}}.}
\format{
A data frame with 248 rows and 2 columns.
}
\source{
\url{https://datahub.io/core/country-list}
}
\usage{
ifs_countries
}
\description{
ifs_countries
A data frame containing the name of 248 countries and its ISO 3166-1 alpha-2 codes
One can use this data frame to search for the correct ISO code of the country he wants do download data using \code{\link{ifs_data}}.
}
\keyword{datasets}
