% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs_data.R
\name{ifs_data}
\alias{ifs_data}
\title{ifs_data}
\usage{
ifs_data(indicator, country, start, end, freq)
}
\arguments{
\item{indicator}{A character, the IFS database indicator as in \code{\link{ifs_indicators}}.}

\item{country}{A character, the ISO code of the country as in \code{\link{ifs_countries}}.}

\item{start}{A number, the year for which you would like to start gathering the data.}

\item{end}{A number, the year for which you would like to start gathering the data.}

\item{freq}{A character, indicating the series frequency. With 'A' for annual, 'Q' for quarterly, and 'M' for monthly.

Please see also \code{\link[imfr]{imf_data}} for a more throughout explanation.}
}
\description{
\code{ifs_data} downloads data from IFS.
A simplification of \code{\link[imfr]{imf_data}}, which sets \code{database_id} to \code{"IFS"}, \code{times} to \code{3} and both \code{return_raw} and \code{print_url} to \code{FALSE}.
}
\examples{
ifs_data(indicator = "ENDE_XDC_USD_RATE", country = c("UK","BR"), start = 1999, end = 2020, freq = "M")
}
