% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifs_indicators.R
\docType{data}
\name{ifs_indicators}
\alias{ifs_indicators}
\title{ifs_indicatora
A data frame containing the name and description of all indicators avaible for download at the IFS.
One can use this data frame to search for the correct code of the indicator he wants do download data using \code{\link{ifs_data}}.}
\format{
A data frame with 1668 rows and 2 columns.
}
\usage{
ifs_indicators
}
\description{
ifs_indicatora
A data frame containing the name and description of all indicators avaible for download at the IFS.
One can use this data frame to search for the correct code of the indicator he wants do download data using \code{\link{ifs_data}}.
}
\keyword{datasets}
