% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_converter.R
\name{iso_converter}
\alias{iso_converter}
\title{iso_converter}
\usage{
iso_converter(names)
}
\arguments{
\item{names}{a vector of strings}
}
\description{
This function converts English countries names to their respective ISO codes, using \code{\link{ifs_countries}}.
}
\examples{
iso_converter(names = c("Brazil","Canada"))
}
